remove old dmhelper-local test
